Deleted the "base/package.json" and "base/yarn.lock" files.
The "package.json" file contained code that was used to run production scripts. Among others there was a script compiling styles. The styles of the project in accordance to the changes to the vanilla JavaScript and React versions will no longer be compiled from SCSS but will be directly managed as a single monolithic already minified CSS code inside a JavaScript variable (so as to make it readily accessible to the styles injection service).
There was additionally a script that run production scripts of both v2 and v3 versions, but it was using yarn so it have to rewritten, and it will probably be done in the project main "package.json" file.

The package manager of the project is being switched to npm as I deem it as a more reliable product.