Removed the content of the "base" directory to the main directory.
The "base/index.html" file was moved now to "demo/index.html" instead as webpack dev server can't work properly when the dev server is set up for the parent directory (even though it works if it's a sibling directory as it is now—"demo" is a sibling of "v2" and "v3").

Additionally, all images despite the image named "1" were removed—only one image will be used in the demo.