Added a production file copying command to the v2 version's "package.json" file's production script.
The command is copying the production "package/index.js" file to the "package/v2.js" file. As Vue 2 will soon become obsolete, the "index.js" will become a Vue 3 script and the Vue 2 version will have to be accessed at "fslightbox-vue/v2". Now a transitory period is introduced where the "index.js" file is still pointing to the Vue 2 version, but in the documentation it will be written that the version has to be accessed at the "fslightbox-vue/v2".